requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
cmake==3.27.5
coloredlogs==15.0.1
cssselect==1.2.0
diskcache==5.6.3
exceptiongroup==1.1.3
fake-useragent==1.2.1
fastapi==0.103.2
feedparser==6.0.10
filelock==3.12.4
FLAML==2.1.0
flatbuffers==23.5.26
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
html5lib==1.1
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
lit==17.0.1
lxml==4.9.3
MarkupSafe==2.1.3
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
networkx==3.1
nltk==3.8.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.16.0
openai==0.28.1
overrides==7.4.0
packaging==23.1
pandas==2.1.1
parse==1.19.1
peewee==3.16.3
Pillow==10.0.1
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pyautogen==0.1.2
pydantic==2.4.2
pydantic_core==2.10.1
pyee==8.2.2
PyPika==0.48.9
pyppeteer==1.0.2
pyquery==2.0.0
pysqlite3-binary==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-html==0.10.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
w3lib==2.1.2
watchfiles==0.20.0
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yarl==1.9.2
yfinance==0.2.30
zipp==3.17.0
